tp5/generador_pseudoaliatorio/test/tabla_test.py
import unittest
import sys
sys.path.append('../')
from generador_pseudoaliatorio.tabla import Tabla, Uniforme, Exponencial, Normal, Poisson
import generador_pseudoaliatorio.estadistica as estadistica

# Muestras
muestra_1 = [0.15, 0.22, 0.41, 0.65, 0.84, 0.81, 0.62, 0.45, 0.32, 0.07, 0.11, 0.29, 0.58, 0.73, 0.93, 0.97, 0.79, 0.55,
             0.35, 0.09, 0.99, 0.51, 0.35, 0.02, 0.19, 0.24, 0.98, 0.10, 0.31, 0.17]
muestra_2 = [0.10, 0.25, 1.53, 2.83, 3.50, 4.14, 5.65, 6.96, 7.19, 8.25, 1.20, 5.24, 4.75, 3.96, 2.21, 3.15, 2.53, 1.16,
             0.32, 0.90, 0.87, 1.34, 1.87, 2.91, 0.71, 1.69, 0.69, 0.55, 0.43, 0.26]
muestra_3 = [1.56,2.21,3.15,4.61, 4.18,5.20,6.94,7.71,5.15,6.76,7.28,4.23,3.21,2.75,
            4.69,5.86,6.25,4.27,4.91,4.78,2.46,3.97,5.71,6.19,4.20,3.48,5.83,6.36,5.90,5.43]
muestra_4 = [14,7,13,16,16,13,14,17,15,16,13,15,10,15,16,14,12,17,14,12,13,20,8,17,19,11,12,17,9,18,20,10,18,15,13,16,24,
            18,16,18,12,14,20,15,10,13,21,23,15,18,]


class TestChiUniforme(unittest.TestCase):
    @classmethod
    def setUpClass(cls) -> None:
        cls.tabla = Uniforme(num_intervalos=5, datos=muestra_1, valor_minimo=0, valor_maximo=1, decimals=2)
        cls.tabla.chi()
        cls.fe = fe = []
        cls.fo = fo = []
        cls.c = c = []
        for interv in cls.tabla.intervalos:
            fe.append(interv.fe)
            fo.append(interv.fo)
            c.append(round(interv.c, 15))

    def test_c_acum(self):
        self.assertEqual(1.66666666666667, round(self.tabla.c_acum, 14))

    def test_fe(self):
        self.assertEqual(len(self.tabla.datos), sum(self.fe))
        self.assertListEqual([6, 6, 6, 6, 6], self.fe)

    def test_fo(self):
        self.assertEqual(len(self.tabla.datos), sum(self.fo))
        self.assertListEqual([8, 7, 5, 4, 6], self.fo)

    def test_c(self):
        self.assertListEqual([0.666666666666667, 0.166666666666667, 0.166666666666667, 0.666666666666667, 0], self.c)


class TestChiExponencial(unittest.TestCase):
    @classmethod
    def setUpClass(cls) -> None:
        cls.tabla = Exponencial(num_intervalos=10, datos=muestra_2, valor_minimo=0, valor_maximo=10, decimals=2)
        cls.tabla.chi()
        cls.fe = fe = []
        cls.fo = fo = []
        # Para los intervalos reorganizados
        cls.fe_reorg = fe_reorg = []
        cls.fo_reorg = fo_reorg = []
        cls.c = c = []
        for interv in cls.tabla.intervalos:
            fe.append(round(interv.fe, 5))
            fo.append(interv.fo)

        for interv in cls.tabla.intervalos_reorganizados:
            fe_reorg.append(round(interv.fe, 5))
            fo_reorg.append(interv.fo)
            c.append(round(interv.c, 5))

    def test_media(self):
        self.assertEqual(2.57133333333333, round(estadistica.media(self.tabla.datos), 14))

    def test_lambda(self):
        self.assertEqual(0.388903292714545, round(self.tabla.get_lambda(),15))

    def test_fe(self):
        self.assertListEqual([9.58677, 6.49791, 4.40428, 2.98522, 2.02338, 1.37145,
                              0.92957, 0.63006, 0.42706, 0.28946], self.fe)

    def test_fo(self):
        self.assertListEqual([10, 6, 4, 3, 2, 2, 1, 1, 1, 0], self.fo)

    # Intervalos reorganizados
    def test_fe_reorg(self):
        self.assertListEqual([9.58677, 6.49791, 7.38951, 5.67097], self.fe_reorg)

    def test_fo_reorg(self):
        self.assertListEqual([10, 6, 7, 7], self.fo_reorg)

    def test_c(self):
        self.assertListEqual([0.01781, 0.03815, 0.02053, 0.31147], self.c)

    def test_c_acum(self):
        self.assertEqual(0.3880, round(self.tabla.c_acum, 4))



class TestChiNormal(unittest.TestCase):
    @classmethod
    def setUpClass(cls) -> None:
        cls.tabla = Normal(datos = muestra_3, num_intervalos = 10, valor_minimo=0, valor_maximo=10, decimals=2)
        cls.tabla.chi()
        cls.fe = fe = []
        cls.fo = fo = []
        # Para los intervalos reorganizados
        cls.fe_reorg = fe_reorg = []
        cls.fo_reorg = fo_reorg = []
        cls.c = c = []
        for interv in cls.tabla.intervalos:
            fe.append(round(interv.fe, 5))
            fo.append(interv.fo)

        for interv in cls.tabla.intervalos_reorganizados:
            fe_reorg.append(round(interv.fe, 5))
            fo_reorg.append(interv.fo)
            c.append(round(interv.c, 5))

    #Probando los datos del Excel contra los de nuestro modelo
    def test_media(self):
        self.assertEqual(4.8410, round(estadistica.media(self.tabla.datos), 4))


    def test_varianza(self):
        self.assertEqual(2.43, round(estadistica.varianza(self.tabla.datos), 2))

    def test_desviacion(self):
        self.assertEqual(1.5574, round(estadistica.desviacion(self.tabla.datos),4))

    def test_fe(self):
        #Hay diferencias decimales que no son por redondeo
        self.assertListEqual([0.15653,0.76432,2.47108,5.28976,7.49761,7.03635,4.37229,1.79891,0.49006,0.08839], self.fe)

    def test_fo(self):
        self.assertListEqual([0,1,3,4,8,7,5,2,0,0], self.fo)

    # Intervalos reorganizados
    def test_fe_reorg(self):
        # RT al anterior
        self.assertListEqual([8.68170,7.49761,7.03635,6.74965], self.fe_reorg)

    def test_fo_reorg(self):
        self.assertListEqual([8,8,7,7], self.fo_reorg)

    def test_c(self):
        # Como los Fe estan man su correspondiente C va a estar mal
        self.assertListEqual([0.05353,0.03366,0.00019,0.00929], self.c)

    def test_c_acum(self):
        # Idem a anterior, C estan mal
        self.assertEqual(0.0967, round(self.tabla.c_acum, 4))


#Falta terminar
# class TestChiPoisson(unittest.TestCase):
#     @classmethod
#     def setUpClass(cls) -> None:
#         cls.tabla = Poisson(datos = muestra_4, decimals=2)
#         cls.tabla.chi()
#         cls.intervalos = intervalos = []
#         cls.intervalos_reorganizados = intervalos_reorganizados = []
#         cls.fe = fe = []
#         cls.fo = fo = []
#         # Para los intervalos reorganizados
#         cls.fe_reorg = fe_reorg = []
#         cls.fo_reorg = fo_reorg = []
#         cls.c = c = []
#         #Redondeo los valores de Fe para cada intervalo
#         #Prq al ser Poisson no uso decimales
#         for interv in cls.tabla.intervalos:
#             intervalos.append(interv.inicio)
#             fe.append(round(interv.fe, 0))
#             fo.append(interv.fo)
#         for interv in cls.tabla.intervalos_reorganizados:
#             intervalos_reorganizados.append(interv.fin)
#             fe_reorg.append(round(interv.fe, 0))
#             fo_reorg.append(interv.fo)
#             c.append(round(interv.c, 5))
#
#     #Probando los datos del Excel contra los de nuestro modelo
#     #En caso de lamda el valor con decimales es de 15.04 y es correcto
#     #Pero en el excel calcula la densidad con 15 provocando diferencias
#     #En el calculo del fe
#     def test_lamda(self):
#         self.assertEqual(15, self.tabla.get_lambda())
#
#     #Comprobar que los intervalos estan bien armados
#     def test_intervalos(self):
#         self.assertListEqual([7,8,9,10,11,12,13,14,15,16,17,18,19,20,21,22,23,24],self.intervalos)
#
#     #Testeo que cuente bien los datos de cada intervalo
#     def test_fo(self):
#         self.assertListEqual([1,1,1,3,1,4,6,5,6,6,4,5,1,3,1,0,1,1], self.fo)
#     def test_fe(self):
#         self.assertListEqual([1,1,2,2,3,4,5,5,5,5,4,4,3,2,1,1,1,0], self.fe)
#
#     # Compara el FIN de intervalos reorganizados
#     def test_intervalos_reorg(self):
#         pass
#         #self.assertListEqual([10,12,13,14,15,16,18,24], self.intervalos_reorganizados)
#
#     # El resto de los valores van a estar mal por ende
#     def test_fo_reorg(self):
#         pass
#         #self.assertListEqual([6,5,6,5,6,6,9,7], self.fo_reorg)
#
#     def test_fe_reorg(self):
#         pass
#         #self.assertListEqual([6,7,5,5,5,5,8,8], self.fe_reorg)
#
#     def test_c(self):
#         pass
#         # Como los Fe estan man su correspondiente C va a estar mal
#         #self.assertListEqual([0.0000,0.5714,0.2000,0.0000,0.2000,0.2000,0.1250,0.1250], self.c)
#
#     def test_c_acum(self):
#         pass
#         # Idem a anterior, C estan mal
#         #self.assertEqual(1.4214, round(self.tabla.c_acum, 4))